Add antibody target filter and filter options, move get_all_filter_options function to data provider service
