+++ app/database_etl/postgres_tables_handler/__init__.py
@@ -2,4 +2,4 @@ from .table_generator import create_multi_select_tables, create_bridge_tables, c
     create_dashboard_source_df, create_research_source_df
 from .table_formatter import format_dashboard_source, add_mapped_variables
 from .postgres_loader import validate_records, load_postgres_tables, drop_table_entries
-from .postgres_utils import get_all_filter_options, check_filter_options, validate_pooling_function_columns
+from .postgres_utils import get_filter_static_options, check_filter_options, validate_pooling_function_columns
