+++ app/namespaces/data_provider/data_provider_controller.py
@@ -3,10 +3,9 @@ import logging.config
 from flask_restplus import Resource, Namespace
 from flask import jsonify, make_response, request
 
-from .data_provider_service import get_record_details, get_country_seroprev_summaries, jitter_pins
+from .data_provider_service import get_record_details, get_country_seroprev_summaries, jitter_pins, get_all_filter_options
 from .data_provider_schema import RecordDetailsSchema, RecordsSchema, PaginatedRecordsSchema, StudyCountSchema
 from app.utils import validate_request_input_against_schema, get_filtered_records, get_paginated_records, convert_start_end_dates
-from app.database_etl.postgres_tables_handler import get_all_filter_options
 
 data_provider_ns = Namespace('data_provider', description='Endpoints for getting database records.')
 logging.getLogger(__name__)
